Fix petastorm/__init__.py to not include predicates/selectors so pip install works. Added development instructions to README.
